// src/libs/renderer/styles.css.d.ts
declare const styles: {
  p: string;
  heading: string;
  heading_button: string;
  heading_button_level: string;
  ordered_list_item: string;
  ordered_list_item_number: string;
  ordered_list_item_dot: string;
  unordered_list_item: string;
  unordered_list_item_dot: string;
  todo_item: string;
  checkbox: string;
  checkbox_svg: string;
  checkbox_background: string;
  icon: string;
  todo_item_done: string;
  blockquote: string;
  blockquote_markup: string;
  hr: string;
  code_block: string;
  inline_markup: string;
  code_language: string;
  code_close: string;
  code_span: string;
  code_span_inner: string;
  code_span_open: string;
  code_span_close: string;
  reference: string;
  mark: string;
  mark_markup: string;
  strikethrough: string;
  underline: string;
  tag: string;
  tag_markup: string;
  link_button: string;
  link: string;
  link_open: string;
  link_close: string;
  link_nowrap: string;
  image: string;
  file: string;
  file_svg: string;
  [key: string]: string;
};

export default styles;